+++ apps/docs/src/app/layout.tsx
@@ -27,7 +27,7 @@ const RootLayout: FC<PropsWithChildren> = async (props) => {
     <html lang="de" suppressHydrationWarning>
       <body className={bodyClassName}>
         <div className={styles.background} />
-        <ThemeProvider attribute="data-flow-theme" disableTransitionOnChange>
+        <ThemeProvider attribute="data-theme" disableTransitionOnChange>
           <RouterProvider>
             <ScrollToHash />
             <Matomo />