+++ packages/components/src/components/List/components/Items/Items.tsx
@@ -6,7 +6,6 @@ import clsx from "clsx";
 import type { FC } from "react";
 import styles from "./Items.module.scss";
 import { FallbackItems } from "./components/FallbackItems";
-import ListEmptyViewContainerView from "@/views/ListEmptyViewContainerView";
 
 export const Items: FC = () => {
   const list = useList();
@@ -33,13 +32,6 @@ export const Items: FC = () => {
       <ItemsGridListView
         className={rootClassName}
         {...list.componentProps}
-        emptyView={
-          <ListEmptyViewContainerView
-            viewType={list.getEmptyViewType()}
-            emptySearchResultView={list.emptySearchResultView}
-            emptyView={list.emptyView}
-          />
-        }
         layout={tiles ? "grid" : "stack"}
         tileMaxWidth={list.itemView.tileMaxWidth}
       >